Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
jaeger-client==4.3.0
prometheus-client==0.12.0
opentelemetry-api==0.12b0
opentelemetry-exporter-jaeger==0.12b0
opentelemetry-instrumentation-flask==0.12b0
opentelemetry-instrumentation-requests==0.12b0
opentelemetry-instrumentation==0.12b0
opentelemetry-sdk==0.12b0
prometheus-flask-exporter==0.18.1
pylint==2.4.0
requests==2.23.0
pymongo==3.11.0
Flask-PyMongo==2.3.0

MarkupSafe==2.0.1
jinja2==3.0.3
itsdangerous==2.0.1
werkzeug==2.0.3
Flask-OpenTracing